# Conflicts:
#	Backend/Social/EventHandler/IntegrationEvents/UserEvents/CreatedReviewEventHandler.cs
#	Backend/SocialFeed/Domain/Post.cs
#	Backend/SocialFeed/Infrastructure/PostRepository.cs

+++ Frontend/Entities/User.cs
@@ -15,7 +15,7 @@ public class User
     
     public int RatedMovies { get; set; }
     
-    public float AverageOfUserRatings { get; set; }
+    public double AverageOfUserRatings { get; set; }
     
     
 }
